fix(poc): address high-sev RLCC API spec gaps in receiver/assembler
Audit of poc/backend against RLCC API Documentation v0.1 (§4-§6) surfaced
eight high-severity bugs. This commit fixes six of them:

- BillReprint now reads `billNumber` (spec key) with fallback to legacy
  `transactionNumber`; new `_parse_ts_or_ms` accepts both ISO 8601 and
  Long-ms `transactionTimestamp` so the alert lands at the right time.
- BeginTransactionWithTillLookup returns the spec envelope
  `{data: {ErrorCode, Succeeded, TransactionSessionId}}` instead of the
  generic `{message, event}` — Nukkad's client can now thread the session
  id back into subsequent calls.
- GetTill performs an actual lookup against camera_mapping (matches
  storeIdentifier + tillDescription against display_pos_label /
  pos_terminal_no) and returns the spec envelope on hit, 400 +
  ErrorCode 11 on miss.
- SaleLine.item_quantity is now `float`; `int(itemQuantity)` rejected
  fractional kg quantities like "1.250" with a ValueError → 400.
- New `_truthy()` parses booleans strictly (rejects the literal string
  "false"); employeePurchase and isPreviousTransaction were silently
  flipping to True whenever Nukkad sent stringified-JSON bodies.

Also expands `verify_push_endpoints.py`:
- happy_path now asserts the Begin envelope shape.
- New scenarios: fractional_qty (1.250 kg), stringified_bool ("false"
  strings), get_till_unknown (asserts 400 + ErrorCode 11).
- Renamed get_till → get_till_unknown to reflect what it actually checks.

Not addressed in this pass (intentional — see audit notes):
- Enum validation across §6.1-§6.8 (still pass-through; best added at
  the _ingest() boundary in a follow-up).
- Dropped optional keys (branch/tillDescription/currencyCode/printable/
  *LinkedTo chain) — useful for richer fraud detection later, none
  blocking.
- applicationType not validated against "Retail".

+++ poc/backend/models.py
@@ -20,7 +20,7 @@ class SaleLine(BaseModel):
     line_number: int = 0
     item_id: str = ""
     item_description: str = ""
-    item_quantity: int = 0
+    item_quantity: float = 0.0
     item_unit_price: float = 0.0
     total_amount: float = 0.0
     scan_attribute: str = "None"
@@ -36,7 +36,9 @@ class SaleLine(BaseModel):
             line_number=payload.get("lineNumber", 0),
             item_id=payload.get("itemID", ""),
             item_description=payload.get("itemDescription", ""),
-            item_quantity=int(payload.get("itemQuantity", 0) or 0),
+            # Float, not int: spec marks itemQuantity as integer but real grocery
+            # POS systems send fractional kg (e.g. "1.250").
+            item_quantity=float(payload.get("itemQuantity", 0) or 0),
             item_unit_price=float(payload.get("itemUnitPrice", 0.0) or 0.0),
             total_amount=float(payload.get("totalAmount", 0.0) or 0.0),
             scan_attribute=payload.get("scanAttribute", "None"),
